refactor(order): 주문 애그리거트 패키지 구조 변경
