+++ src/main/java/com/musinsa/orders/domain/order/Money.java
@@ -1,4 +1,4 @@
-package com.musinsa.orders.domain;
+package com.musinsa.orders.domain.order;
 
 import java.util.Objects;
 import javax.persistence.Embeddable;
