+++ src/main/java/com/musinsa/orders/domain/order/ShippingFeePolicy.java
@@ -1,4 +1,4 @@
-package com.musinsa.orders.domain;
+package com.musinsa.orders.domain.order;
 
 public interface ShippingFeePolicy {
