+++ src/main/java/com/musinsa/orders/infra/order/JpaOrderRepository.java
@@ -1,7 +1,7 @@
-package com.musinsa.orders.infra;
+package com.musinsa.orders.infra.order;
 
-import com.musinsa.orders.domain.Order;
-import com.musinsa.orders.domain.OrderRepository;
+import com.musinsa.orders.domain.order.Order;
+import com.musinsa.orders.domain.order.OrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaOrderRepository extends OrderRepository, JpaRepository<Order, Long> {
