+++ src/test/java/com/musinsa/orders/domain/ExchangePolicyTest.java
@@ -6,6 +6,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 import com.musinsa.orders.domain.exchange.ExchangePolicy;
+import com.musinsa.orders.domain.order.Money;
+import com.musinsa.orders.domain.order.Order;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
