+++ src/test/java/com/musinsa/orders/domain/MoneyTest.java
@@ -3,6 +3,7 @@ package com.musinsa.orders.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
+import com.musinsa.orders.domain.order.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
