+++ src/test/java/com/musinsa/orders/domain/RefundPolicyTest.java
@@ -4,6 +4,8 @@ import static com.musinsa.orders.Fixtures.createOrder;
 import static com.musinsa.orders.Fixtures.createOrderLineItem;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.musinsa.orders.domain.order.Money;
+import com.musinsa.orders.domain.order.Order;
 import com.musinsa.orders.domain.refund.RefundPolicy;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
