+++ src/test/java/com/musinsa/orders/domain/ShippingFeePolicyTest.java
@@ -4,7 +4,10 @@ import static com.musinsa.orders.Fixtures.createOrder;
 import static com.musinsa.orders.Fixtures.createOrderLineItem;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.musinsa.orders.infra.AmountShippingFeePolicy;
+import com.musinsa.orders.domain.order.Money;
+import com.musinsa.orders.domain.order.Order;
+import com.musinsa.orders.domain.order.ShippingFeePolicy;
+import com.musinsa.orders.infra.order.AmountShippingFeePolicy;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
